requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-google-genai
faiss-cpu
sentence-transformers
boto3
python-dotenv
pypdf
docx2txt
